+++ tests/Integration/Models/PlanSubscriptionTest.php
@@ -4,10 +4,10 @@
 
 namespace Gerarodjbaez\Laraplans\Tests\Integration\Models;
 
-use Config;
 use Carbon\Carbon;
 use Gerardojbaez\Laraplans\Period;
 use Illuminate\Support\Facades\Event;
+use Illuminate\Support\Facades\Config;
 use Gerardojbaez\Laraplans\Models\Plan;
 use Gerardojbaez\Laraplans\Tests\TestCase;
 use Gerardojbaez\Laraplans\Tests\Models\User;
@@ -30,7 +30,7 @@ class PlanSubscriptionTest extends TestCase
      *
      * @return  void
      */
-    public function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
 
@@ -82,7 +82,7 @@ class PlanSubscriptionTest extends TestCase
         $subscription = factory(PlanSubscription::class)->create();
 
         // Assert
-        Event::assertFired(SubscriptionCreated::class, function ($event) use ($subscription) {
+        Event::assertDispatched(SubscriptionCreated::class, function ($event) use ($subscription) {
             return (int) $event->subscription->id === (int) $subscription->id;
         });
     }
@@ -139,7 +139,7 @@ class PlanSubscriptionTest extends TestCase
 
         $subscription = $this->subscription;
 
-        Event::assertFired(SubscriptionCanceled::class, function ($event) use ($subscription) {
+        Event::assertDispatched(SubscriptionCanceled::class, function ($event) use ($subscription) {
             return (int) $event->subscription->id === (int) $subscription->id;
         });
     }
@@ -189,7 +189,7 @@ class PlanSubscriptionTest extends TestCase
         $this->assertTrue($subscription->isActive());
         $this->assertEquals(Carbon::now()->addMonth()->month, $subscription->ends_at->month);
 
-        Event::assertFired(SubscriptionRenewed::class, function ($event) use ($subscription) {
+        Event::assertDispatched(SubscriptionRenewed::class, function ($event) use ($subscription) {
             return (int) $event->subscription->id === (int) $subscription->id;
         });
     }
@@ -291,6 +291,58 @@ class PlanSubscriptionTest extends TestCase
     }
 
     /**
+     * Exclude canceled subscription.
+     *
+     * @test
+     * @return void
+     */
+    public function it_can_exclude_canceled_subscriptions()
+    {
+        // Make sure canceled subscriptions are not included...
+        factory(PlanSubscription::class, 10)->create([
+            'canceled_at' => Carbon::now()->subDays(2)
+        ]);
+
+        // Not canceled subscriptions
+        factory(PlanSubscription::class, 5)->create([
+            'canceled_at' => null,
+        ]);
+
+        $result = PlanSubscription::excludeCanceled()->get();
+
+        $this->assertEquals(6, $result->count());
+    }
+
+    /**
+     * Exclude immediately canceled subscription.
+     *
+     * @test
+     * @return void
+     */
+    public function it_can_exclude_immediately_canceled_subscriptions()
+    {
+        // Immediately canceled subscriptions...
+        factory(PlanSubscription::class, 2)->create([
+            'canceled_at' => Carbon::now(),
+            'canceled_immediately' => 1
+        ]);
+
+        // Not immediately canceled subscriptions...
+        factory(PlanSubscription::class, 3)->create([
+            'canceled_at' => null,
+        ]);
+
+        factory(PlanSubscription::class, 2)->create([
+            'canceled_at' => null,
+            'canceled_immediately' => 0
+        ]);
+
+        $result = PlanSubscription::excludeImmediatelyCanceled()->get();
+
+        $this->assertEquals(6, $result->count());
+    }
+
+    /**
      * Can change subscription plan.
      *
      * @test
@@ -298,7 +350,9 @@ class PlanSubscriptionTest extends TestCase
      */
     public function it_can_change_plan()
     {
-        Event::fake();
+        Event::fake([
+            SubscriptionPlanChanged::class
+        ]);
 
         $newPlan = Plan::create([
             'name' => 'Business',
@@ -336,7 +390,7 @@ class PlanSubscriptionTest extends TestCase
 
         $subscription = $this->subscription;
 
-        Event::assertFired(SubscriptionPlanChanged::class, function ($event) use ($subscription) {
+        Event::assertDispatched(SubscriptionPlanChanged::class, function ($event) use ($subscription) {
             return (int) $event->subscription->id === (int) $subscription->id;
         });
     }